Abstracted useAuth hook to separate folder
